app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
boto3>=1.34.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.4.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
